protocol client abstraction; separated the protocol client sending requests and getting responses from the interaction affordances implemented by the ConsumedThing
