+++ WoT/core/TDHelpers.cs
